Move resource commands to own namespaces
