+++ CommandLine/src/Commands/Resources/Delete.cs
@@ -1,13 +1,14 @@
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
+using Worms.Commands.Resources.Schemes;
 
 // ReSharper disable UnusedMember.Global - CLI library uses magic to call OnExecuteAsync()
 
-namespace Worms.Commands
+namespace Worms.Commands.Resources
 {
-    [Command("create", Description = "Create a resource")]
-    [Subcommand(typeof(CreateScheme))]
-    internal class Create : CommandBase
+    [Command("delete", "rm", Description = "Delete a resource")]
+    [Subcommand(typeof(DeleteScheme))]
+    internal class Delete : CommandBase
     {
         public Task<int> OnExecuteAsync(CommandLineApplication app)
         {
