+++ CommandLine/src/Commands/Resources/Get.cs
@@ -1,9 +1,11 @@
 using System.Threading.Tasks;
 using McMaster.Extensions.CommandLineUtils;
+using Worms.Commands.Resources.Games;
+using Worms.Commands.Resources.Schemes;
 
 // ReSharper disable UnusedMember.Global - CLI library uses magic to call OnExecuteAsync()
 
-namespace Worms.Commands
+namespace Worms.Commands.Resources
 {
     [Command("get", Description = "Get a list of resources")]
     [Subcommand(typeof(GetScheme))]
